requirements.txt
pygame
websockets
websocket-client
tinytag
rich
watchdog
yt-dlp
paramiko
scp
thefuzz
pynput
scipy
pandas
tqdm
watchdog
pigpio
rich
pyserial
git+https://github.com/aubio/aubio.git